requirements.txt
pandas
python-docx
sqlalchemy
openai
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
pyprojroot
openpyxl


weaviate-client[agents]
pandas
python-docx
openpyxl
python-dotenv
streamlit
weaviate-client